Feat: 회원 탈퇴 프론트

+++ src/user/user.service.ts
@@ -60,7 +60,9 @@ export class UserService {
   }
 
   async findUserByEmail(email: string) {
-    return await this.userRepository.findOne({ where: { email } });
+    const user = await this.userRepository.findOne({ where: { email } });
+    if (!user) throw new NotFoundException();
+    return user;
   }
 
   async softDeleteUser(id: number) {
@@ -69,14 +71,16 @@ export class UserService {
       relations: ['boards'],
     });
 
-    if (!existingUser)
+    if (!existingUser) {
       throw new NotFoundException('id와 일치하는 유저가 없습니다.');
-
+    }
     try {
       await this.userRepository.softDelete(id);
-      await this.boardRepository.delete(
-        existingUser.boards.map((board) => board.id),
-      );
+      if (existingUser.boards.length) {
+        await this.boardRepository.delete(
+          existingUser.boards.map((board) => board.id),
+        );
+      }
       return { message: '회원탈퇴를 완료했습니다.' };
     } catch (error) {
       throw new HttpException('서버 에러', HttpStatus.INTERNAL_SERVER_ERROR);
